Working on removing the DYLD_LIBRARY_PATH setting
setting this variable has undesireable side effects and effects all programs that are run in this environment.
Instead of using this method all linking should be handled by setting the rpath and using that in compilation.
This method helps ensure using this program does not break anything for anyone else

Additionally, working on experimenting a bit with adding support for units